Deleted shared.ts, all functions moved to utilities.ts
