+++ test/smoke/src/helpers/utilities.ts
@@ -3,6 +3,8 @@
 
 import * as fs from "fs";
 import * as cp from "child_process";
+import * as request from "request";
+import * as URL from "url-parse";
 import { dirname } from "path";
 import { SpawnSyncOptions } from "child_process";
 
@@ -63,4 +65,55 @@ export function spawnSync(command: string, args?: string[], options?: SpawnSyncO
     if (result.error) {
         throw result.error;
     }
+}
+
+export function getContents(url, token, headers, callback) {
+    request.get(toRequestOptions(url, token, headers), function (error, response, body) {
+        if (!error && response && response.statusCode >= 400) {
+            error = new Error("Request returned status code: " + response.statusCode + "\nDetails: " + response.body);
+        }
+
+        callback(error, body);
+    });
+}
+
+export function toRequestOptions(url, token?, headers?) {
+    headers = headers || {
+        "user-agent": "nodejs",
+    };
+
+    if (token) {
+        headers["Authorization"] = "token " + token;
+    }
+
+    let parsedUrl = new URL(url);
+
+    let options: any = {
+        url: url,
+        headers: headers,
+    };
+
+    // We need to test the absence of true here because there is an npm bug that will not set boolean
+    // env variables if they are set to false.
+    if (process.env.npm_config_strict_ssl !== "true") {
+        options.strictSSL = false;
+    }
+
+    if (process.env.npm_config_proxy && parsedUrl.protocol === "http:") {
+        options.proxy = process.env.npm_config_proxy;
+    } else if (process.env.npm_config_https_proxy && parsedUrl.protocol === "https:") {
+        options.proxy = process.env.npm_config_https_proxy;
+    }
+
+    return options;
+}
+
+// Await function
+export async function sleep(time: number) {
+    await new Promise(resolve => {
+        const timer = setTimeout(() => {
+            clearTimeout(timer);
+            resolve();
+        }, time);
+    });
 }
